Renaming of whatsapp module
